+++ Assets/Editor Toolbox/Editor/Drawers/ToolboxDrawers/PropertyCondition.cs
@@ -1,0 +1,9 @@
+﻿namespace Toolbox.Editor.Drawers
+{
+    public enum PropertyCondition
+    {
+        Valid,
+        NonValid,
+        Disabled
+    }
+}