+++ Assets/Editor Toolbox/Editor/Drawers/ToolboxDrawers/ToolboxListPropertyDrawer.cs
@@ -7,7 +7,6 @@ namespace Toolbox.Editor.Drawers
     {
         public override bool IsPropertyValid(SerializedProperty property)
         {
-            //NOTE: this will be always true since the ToolboxPropertyHandler will validate each drawer
             return property.isArray;
         } 
     }